requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
dateparser==1.3.0
frozenlist==1.8.0
idna==3.11
loguru==0.7.3
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.1
propcache==0.4.1
pycryptodome==3.23.0
Pygments==2.19.2
python-binance==1.0.34
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
regex==2026.1.15
requests==2.32.5
rich==14.3.2
shellingham==1.5.4
six==1.17.0
typer==0.23.1
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
websockets==16.0
win32_setctime==1.2.0
yarl==1.22.0
# Binance Futures Trading Bot

## Overview
This project is a Python-based Binance Futures trading bot that supports:

- EMA Crossover Strategy
- Risk Management (1% risk per trade)
- Stop Loss and Take Profit
- Backtesting Engine
- Performance Metrics (Win Rate, Drawdown, Profit Factor)

## Technologies Used
- Python
- Binance Futures Connector
- Pandas
- NumPy

## Features
- Fetch live market data
- Execute EMA crossover strategy
- Backtest strategy on historical data
- Risk-based position sizing
- Performance reporting

## Setup Instructions
1. Install dependencies: